+++ test/JsonDocument/StaticJsonDocument.cpp
@@ -33,27 +33,23 @@ TEST_CASE("StaticJsonDocument") {
     StaticJsonDocument<200> doc1, doc2;
     doc1.to<JsonVariant>().set(666);
     deserializeJson(doc2, "{\"hello\":\"world\"}");
-    doc2.nestingLimit = 42;
 
     doc1 = doc2;
 
     std::string json;
     serializeJson(doc1, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc1.nestingLimit == 42);
   }
 
   SECTION("Copy constructor") {
     StaticJsonDocument<200> doc1;
     deserializeJson(doc1, "{\"hello\":\"world\"}");
-    doc1.nestingLimit = 42;
 
     StaticJsonDocument<200> doc2 = doc1;
 
     std::string json;
     serializeJson(doc2, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc2.nestingLimit == 42);
   }
 
   SECTION("Assign from StaticJsonDocument of different capacity") {
@@ -61,14 +57,12 @@ TEST_CASE("StaticJsonDocument") {
     StaticJsonDocument<300> doc2;
     doc1.to<JsonVariant>().set(666);
     deserializeJson(doc2, "{\"hello\":\"world\"}");
-    doc2.nestingLimit = 42;
 
     doc1 = doc2;
 
     std::string json;
     serializeJson(doc1, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc1.nestingLimit == 42);
   }
 
   SECTION("Assign from DynamicJsonDocument") {
@@ -76,39 +70,33 @@ TEST_CASE("StaticJsonDocument") {
     DynamicJsonDocument doc2(4096);
     doc1.to<JsonVariant>().set(666);
     deserializeJson(doc2, "{\"hello\":\"world\"}");
-    doc2.nestingLimit = 42;
 
     doc1 = doc2;
 
     std::string json;
     serializeJson(doc1, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc1.nestingLimit == 42);
   }
 
   SECTION("Construct from StaticJsonDocument of different size") {
     StaticJsonDocument<300> doc2;
     deserializeJson(doc2, "{\"hello\":\"world\"}");
-    doc2.nestingLimit = 42;
 
     StaticJsonDocument<200> doc1 = doc2;
 
     std::string json;
     serializeJson(doc1, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc1.nestingLimit == 42);
   }
 
   SECTION("Construct from DynamicJsonDocument") {
     DynamicJsonDocument doc2(4096);
     deserializeJson(doc2, "{\"hello\":\"world\"}");
-    doc2.nestingLimit = 42;
 
     StaticJsonDocument<200> doc1 = doc2;
 
     std::string json;
     serializeJson(doc1, json);
     REQUIRE(json == "{\"hello\":\"world\"}");
-    REQUIRE(doc1.nestingLimit == 42);
   }
 }